modified:   main.c

+++ src/main.c
@@ -55,14 +55,14 @@ int read_input_array(long long out_array[], int array_max_size, struct interval_
     int num;
 
     do {
-        if(scanf("%i", &num) != 1)
+        if(scanf("%lli", &num) != 1)
             return size;
         if(num <= interval.from && interval.from_set_flag == 1)
-            printf("%lli ", (long long)num);
+            printf("%lli ", num);
         if(num >= interval.to && interval.to_set_flag == 1)
-            stderr_printf("%lli ", (long long)num);
+            stderr_printf("%lli ", num);
         if((num > interval.from || interval.from_set_flag == 0) && (num < interval.to || interval.to_set_flag == 0)) {
-            out_array[size] = (long long)num;
+            out_array[size] = num;
             size++;
         }
     } while (size < array_max_size);
